# swagger-docs/swagger.json
{
    "swagger": "2.0",
    "paths": {
...
        "/orderschedule/index.php/Home/Reassign/preAccept": {
            "get": {
                "tags": [
                    "Reassign"
                ],
                "summary": "\u83b7\u53d6\u5f85\u63a5\u53d7\u6539\u6d3e",
                "description": "",
                "operationId": "",
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "X-Rate-Limit": {
                                "description": "calls per hour allowed by the user",
                                "type": "integer",
                                "format": "int32"
                            },
                            "X-Expires-After": {
                                "description": "date in UTC when token expires",
                                "type": "string",
                                "format": "date-time"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
        "/orderschedule/index.php/Home/Reassign/Reassigning": {
            "get": {
                "tags": [
                    "Reassign"
                ],
                "summary": "\u83b7\u53d6\u6539\u6d3e\u8ba2\u5355",
                "description": "",
                "operationId": "",
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "X-Rate-Limit": {
                                "description": "calls per hour allowed by the user",
                                "type": "integer",
                                "format": "int32"
                            },
                            "X-Expires-After": {
                                "description": "date in UTC when token expires",
                                "type": "string",
                                "format": "date-time"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
        "/orderschedule/index.php/Home/Reassign/accept": {
            "post": {
                "tags": [
                    "Reassign"
                ],
                "summary": "\u63a5\u53d7\u6d3e\u9001",
                "description": "",
                "operationId": "",
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "\u5fc5\u4f20",
                        "description": "\u8ba2\u5355id",
                        "required": true,
                        "type": "int"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "X-Rate-Limit": {
                                "description": "calls per hour allowed by the user",
                                "type": "integer",
                                "format": "int32"
                            },
                            "X-Expires-After": {
                                "description": "date in UTC when token expires",
                                "type": "string",
                                "format": "date-time"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
        "/orderschedule/index.php/Home/Reassign/reject": {
            "post": {
                "tags": [
                    "Reassign"
                ],
                "summary": "\u62d2\u7edd\u6d3e\u9001",
                "description": "",
                "operationId": "",
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "\u5fc5\u4f20",
                        "description": "\u8ba2\u5355id",
                        "required": true,
                        "type": "int"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "X-Rate-Limit": {
                                "description": "calls per hour allowed by the user",
                                "type": "integer",
                                "format": "int32"
                            },
                            "X-Expires-After": {
                                "description": "date in UTC when token expires",
                                "type": "string",
                                "format": "date-time"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
        "/orderschedule/index.php/Home/Reassign/cancel": {
            "post": {
                "tags": [
                    "Reassign"
                ],
                "summary": "\u53d6\u6d88\u6d3e\u9001",
                "description": "",
                "operationId": "",
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "\u5fc5\u4f20",
                        "description": "\u8ba2\u5355id",
                        "required": true,
                        "type": "int"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "X-Rate-Limit": {
                                "description": "calls per hour allowed by the user",
                                "type": "integer",
                                "format": "int32"
                            },
                            "X-Expires-After": {
                                "description": "date in UTC when token expires",
                                "type": "string",
                                "format": "date-time"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
        "/orderschedule/index.php/Home/Reassign/Reassign": {
            "post": {
                "tags": [
                    "Reassign"
                ],
                "summary": "\u521b\u5efa\u6539\u6d3e",
                "description": "",
                "operationId": "",
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "\u5fc5\u4f20",
                        "description": "\u8ba2\u5355id",
                        "required": true,
                        "type": "int"
                    },
                    {
                        "name": "targetId",
                        "in": "\u5fc5\u4f20",
                        "description": "\u76ee\u6807\u6d3e\u9001\u5458id",
                        "required": true,
                        "type": "int"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "X-Rate-Limit": {
                                "description": "calls per hour allowed by the user",
                                "type": "integer",
                                "format": "int32"
                            },
                            "X-Expires-After": {
                                "description": "date in UTC when token expires",
                                "type": "string",
                                "format": "date-time"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
...
    },
...
}